pyproject.toml
[project]
name = "ai-job-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "langchain-core>=1.2.16",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.10",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.22",
]
